Record handle extension now is variant of hash-of-tag if it is a single entry record, or number of nested entries if composite

+++ components/model/src/append_entry.rs
@@ -10,5 +10,5 @@ pub struct AppendEntry {
     pub offset: u64,
 
     /// Quantity of nested records
-    pub len: u32,
+    pub len: u16,
 }
